knowledge-base-service/requirements.txt
fastapi
uvicorn
boto3                       # For AWS S3 storage (future integration)
motor                       # Async driver for MongoDB
pymongo                     # MongoDB integration
google-auth                 # Authentication for Google APIs
google-cloud-storage        # Integration with Google Cloud Storage
dropbox                     # Integration with Dropbox
pydantic[email]             # Extended email validation
pydantic-settings
python-multipart            # File upload support in FastAPI
pytest                      # For testing
pytest-mock
pytest-asyncio              # Async testing with pytest
pycodestyle                 # For code linting and PEP 8 compliance
